Agent-Recommendation/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-multipart==0.0.6
google-generativeai==0.3.1
python-dotenv==1.0.0
scikit-learn==1.3.2
nltk==3.8.1 